OrthoBeam co-planar lines
When two lines are coplanar, a line that is orthogonal to both and that is the single shortest line that joins them can't exist. However, by specifying a point A along one of the lines, a point B can be found along the second line so that a unique line AB with the minimum distance exists. This applies to all co-planar lines, including parallel lines.